pyproject.toml
[project]
name = "consign"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.14",
    "google-api-python-client>=2.174.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "pandas>=2.3.0",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
